third_party/blink/web_tests/external/wpt/css/css-counter-styles/counter-style-name-syntax-expected.txt

This is a testharness.js-based test.
PASS @counter-style name foo is valid
PASS @counter-style name decimal-leading-zero is valid
PASS @counter-style name lower-alpha is valid
PASS @counter-style name cjk-decimal is valid
PASS @counter-style name japanese-informal is valid
PASS @counter-style name ethiopic-numeric is valid
PASS @counter-style name  is invalid
PASS @counter-style name none is invalid
PASS @counter-style name initial is invalid
PASS @counter-style name inherit is invalid
PASS @counter-style name unset is invalid
PASS @counter-style name decimal is invalid
PASS @counter-style name disc is invalid
FAIL @counter-style name square is invalid assert_equals: expected (undefined) undefined but got (object) object "[object CSSCounterStyleRule]"
FAIL @counter-style name circle is invalid assert_equals: expected (undefined) undefined but got (object) object "[object CSSCounterStyleRule]"
FAIL @counter-style name disclosure-open is invalid assert_equals: expected (undefined) undefined but got (object) object "[object CSSCounterStyleRule]"
FAIL @counter-style name disclosure-closed is invalid assert_equals: expected (undefined) undefined but got (object) object "[object CSSCounterStyleRule]"
Harness: the test ran to completion.
